Cut the dead EI Bench: 5968417
