+++ tuner/src/searchtune/mod.rs
@@ -2,7 +2,6 @@ pub mod bounds;
 pub mod cache;
 pub mod cmaes;
 pub mod elo;
-pub mod optimizer;
 pub mod palette;
 pub mod pentanomial;
 pub mod run;
